feat: open proper tab by default and refactor to use existing hook
