+++ packages/twenty-front/src/modules/workflow/workflow-diagram/hooks/useResetWorkflowRunRightDrawerTabIfNeeded.ts
@@ -1,0 +1,71 @@
+import { getIsInputTabDisabled } from '@/command-menu/pages/workflow/step/view-run/utils/getIsInputTabDisabled';
+import { getIsOutputTabDisabled } from '@/command-menu/pages/workflow/step/view-run/utils/getIsOutputTabDisabled';
+import { activeTabIdComponentState } from '@/ui/layout/tab/states/activeTabIdComponentState';
+import { getSnapshotValue } from '@/ui/utilities/recoil-scope/utils/getSnapshotValue';
+import { WorkflowDiagramRunStatus } from '@/workflow/workflow-diagram/types/WorkflowDiagram';
+import { WORKFLOW_RUN_STEP_SIDE_PANEL_TAB_LIST_COMPONENT_ID } from '@/workflow/workflow-steps/constants/WorkflowRunStepSidePanelTabListComponentId';
+import { WorkflowRunTabId } from '@/workflow/workflow-steps/types/WorkflowRunTabId';
+import { isNull } from '@sniptt/guards';
+import { useRecoilCallback } from 'recoil';
+
+export const useResetWorkflowRunRightDrawerTabIfNeeded = () => {
+  const resetWorkflowRunRightDrawerTabIfNeeded = useRecoilCallback(
+    ({ snapshot, set }) =>
+      ({
+        workflowSelectedNode,
+        stepExecutionStatus,
+      }: {
+        workflowSelectedNode: string;
+        stepExecutionStatus: WorkflowDiagramRunStatus;
+      }) => {
+        const activeWorkflowRunRightDrawerTab = getSnapshotValue(
+          snapshot,
+          activeTabIdComponentState.atomFamily({
+            instanceId: WORKFLOW_RUN_STEP_SIDE_PANEL_TAB_LIST_COMPONENT_ID,
+          }),
+        ) as WorkflowRunTabId | null;
+
+        const isInputTabDisabled = getIsInputTabDisabled({
+          stepExecutionStatus,
+          workflowSelectedNode,
+        });
+        const isOutputTabDisabled = getIsOutputTabDisabled({
+          stepExecutionStatus,
+        });
+
+        if (isNull(activeWorkflowRunRightDrawerTab)) {
+          const defaultTabId = isOutputTabDisabled
+            ? WorkflowRunTabId.NODE
+            : WorkflowRunTabId.OUTPUT;
+
+          set(
+            activeTabIdComponentState.atomFamily({
+              instanceId: WORKFLOW_RUN_STEP_SIDE_PANEL_TAB_LIST_COMPONENT_ID,
+            }),
+            defaultTabId,
+          );
+
+          return;
+        }
+
+        if (
+          (isInputTabDisabled &&
+            activeWorkflowRunRightDrawerTab === WorkflowRunTabId.INPUT) ||
+          (isOutputTabDisabled &&
+            activeWorkflowRunRightDrawerTab === WorkflowRunTabId.OUTPUT)
+        ) {
+          set(
+            activeTabIdComponentState.atomFamily({
+              instanceId: WORKFLOW_RUN_STEP_SIDE_PANEL_TAB_LIST_COMPONENT_ID,
+            }),
+            WorkflowRunTabId.NODE,
+          );
+        }
+      },
+    [],
+  );
+
+  return {
+    resetWorkflowRunRightDrawerTabIfNeeded,
+  };
+};
